requirements.txt
amqp==2.4.2
beautifulsoup4==4.7.1
billiard==3.6.0.0
bs4==0.0.1
CacheControl==0.12.5
cachetools==3.1.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
firebase-admin==2.16.0
Flask==1.0.2
google-api-core==1.9.0
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-firestore==0.32.1
google-cloud-storage==1.15.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.9
grpcio==1.20.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.5.0
MarkupSafe==1.1.1
msgpack==0.6.1
Pillow==6.0.0
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pytz==2019.1
requests==2.21.0
rsa==4.0
six==1.12.0
soupsieve==1.9.1
urllib3==1.24.2
vine==1.3.0
Werkzeug==0.15.2
WTForms==2.2.1
